Add files via upload
Note: Code is not complete, I still have to get the sound waves to propagate from the location of sound so that it hits all the speakers on the x-axis. For now, these lines just loop 30 times.

Program asks for input on frequency, location of sound (for modeling purposes), and how many sensors user would like.